+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 * Support `@loader_path` and `$ORIGIN` for `add_rpathdirs()`
 * Improve `set_version("x.x.x", {build = "%Y%m%d%H%M"})` and add build version
 
+### Bugs fixed
+
+* Fix `set_pcxxheader` bug
+
 ## v2.1.6
 
 ### Changes
@@ -349,6 +353,10 @@
 * 改进`cprint()`，支持24位真彩色输出
 * 对`add_rpathdirs()`增加对`@loader_path`和`$ORIGIN`的内置变量支持，提供可迁移动态库加载
 * 改进`set_version("x.x.x", {build = "%Y%m%d%H%M"})` 支持buildversion设置
+
+### Bugs修复
+
+* 修复`set_pcxxheader`编译没有继承flags配置问题
 
 ## v2.1.6